+++ src/__internal__/popover-menu/index.ts
@@ -1,0 +1,10 @@
+export { default as PopoverMenu } from "./popover-menu.component";
+export {
+  MenuItem,
+  MenuItemLeading,
+  MenuItemLabel,
+  MenuItemSubtext,
+  MenuItemDivider,
+  MenuItemHeading,
+} from "./menu-item";
+export type { PopoverMenuProps } from "./popover-menu.component";