+++ src/__internal__/popover-menu/menu-item/menu-item-content/index.ts
@@ -1,0 +1,3 @@
+export { default as MenuItemLeading } from "./menu-item-leading.component";
+export { default as MenuItemLabel } from "./menu-item-label.component";
+export { default as MenuItemSubtext } from "./menu-item-subtext.component";