+++ src/components/date-range/date-range.style.ts
@@ -18,7 +18,6 @@ const StyledDateRange = styled.div.attrs(applyBaseTheme)<StyledDateRangeProps>`
   ${margin}
 
   & ${StyledDateInput} {
-    width: auto;
     display: inline-block;
     vertical-align: ${({ labelsInline }) => (labelsInline ? "top" : "bottom")};
   }